refator-usecase-test
