+++ ExampleMVVM/Mocks/MoviesRepositoryMockError.swift
@@ -1,0 +1,12 @@
+//
+//  MoviesRepositorySuccessTestError.swift
+//  ExampleMVVM
+//  
+//  Created by tomo on 2024/11/28
+//  
+//
+
+
+enum MoviesRepositoryMockError: Error {
+        case failedFetching
+    }
