+++ ExampleMVVM/Stubs/MoviesPage+Stub.swift
@@ -1,0 +1,20 @@
+//
+//  MoviesPageStub.swift
+//  ExampleMVVMTests
+//  
+//  Created by tomo on 2024/11/28
+//  
+//
+
+import Foundation
+
+extension MoviesPage {
+    static let mock: [MoviesPage] = {
+        let page1 = MoviesPage(page: 1, totalPages: 2, movies: [
+            Movie.stub(id: "1", title: "title1", posterPath: "/1", overview: "overview1"),
+            Movie.stub(id: "2", title: "title2", posterPath: "/2", overview: "overview2")])
+        let page2 = MoviesPage(page: 2, totalPages: 2, movies: [
+            Movie.stub(id: "3", title: "title3", posterPath: "/3", overview: "overview3")])
+        return [page1, page2]
+    }()
+}
